requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
lxml==6.0.0
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
xlsxwriter==3.2.5
